bytenite_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai-whisper==20231117
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 